LPS-156599 Move to service.internal to avoid conflicts with the same-named packages in the portal-security-ldap-impl module
