+++ modules/apps/portal-security/portal-security-ldap-service/src/main/java/com/liferay/portal/security/ldap/service/internal/model/listener/UserModelListener.java
@@ -12,13 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.security.ldap.internal.model.listener;
+package com.liferay.portal.security.ldap.service.internal.model.listener;
 
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
-import com.liferay.portal.kernel.model.UserGroup;
-import com.liferay.portal.kernel.security.exportimport.UserGroupImportTransactionThreadLocal;
+import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.security.ldap.service.LDAPServerAttributeRelLocalService;
 
 import org.osgi.service.component.annotations.Component;
@@ -28,22 +27,36 @@ import org.osgi.service.component.annotations.Reference;
  * @author Michael Bowerman
  */
 @Component(immediate = true, service = ModelListener.class)
-public class UserGroupModelListener extends BaseModelListener<UserGroup> {
+public class UserModelListener extends BaseModelListener<User> {
 
 	@Override
-	public void onAfterCreate(UserGroup userGroup)
+	public void onAfterCreate(User user) throws ModelListenerException {
+		long ldapServerId = user.getLdapServerId();
+
+		if (ldapServerId > 0) {
+			try {
+				_ldapServerAttributeRelLocalService.addLDAPServerAttributeRel(
+					ldapServerId, User.class.getName(), user.getUserId());
+			}
+			catch (Exception exception) {
+				throw new ModelListenerException(exception);
+			}
+		}
+	}
+
+	@Override
+	public void onAfterUpdate(User originalUser, User user)
 		throws ModelListenerException {
 
-		long ldapServerId =
-			UserGroupImportTransactionThreadLocal.getLDAPServerId();
+		long ldapServerId = user.getLdapServerId();
 
-		if (ldapServerId !=
-				UserGroupImportTransactionThreadLocal.DEFAULT_LDAP_SERVER_ID) {
+		if ((ldapServerId > 0) &&
+			!_ldapServerAttributeRelLocalService.hasLDAPServerAttributeRel(
+				ldapServerId, User.class.getName(), user.getUserId())) {
 
 			try {
 				_ldapServerAttributeRelLocalService.addLDAPServerAttributeRel(
-					ldapServerId, UserGroup.class.getName(),
-					userGroup.getUserGroupId());
+					ldapServerId, User.class.getName(), user.getUserId());
 			}
 			catch (Exception exception) {
 				throw new ModelListenerException(exception);
